pyproject.toml
[tool.poetry]
name = "river-flows"
version = "0.1.0"
description = "River flow API"
authors = ["Adam Bailey <ab67319@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^2.7.1"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
ipdb = "^0.13.13"
flake8 = "^7.0.0"
httpx = "^0.27.0"
sqlalchemy = "^2.0.29"
sqlalchemy-utils = "^0.41.2"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
isort = "^5.13.2"
pytest = "^8.2.0"
python-dateutil = "^2.9.0.post0"
vcrpy = "^6.0.1"
apscheduler = "^3.10.4"
pandas = "^2.2.3"
beautifulsoup4 = "^4.13.4"
lxml = "^5.4.0"
mypy = "^1.16.1"
ruff = "^0.12.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
